Move legacy code out of inner
