+++ src/cradle/thinknode/caching.h
@@ -1,5 +1,5 @@
-#ifndef CRADLE_INNER_SERVICE_CORE_H
-#define CRADLE_INNER_SERVICE_CORE_H
+#ifndef CRADLE_THINKNODE_CACHING_H
+#define CRADLE_THINKNODE_CACHING_H
 
 #include <optional>
 
@@ -9,10 +9,12 @@
 #include <cradle/inner/caching/immutable.h>
 #include <cradle/inner/caching/immutable/cache.h>
 #include <cradle/inner/introspection/tasklet.h>
-#include <cradle/inner/service/disk_cached.h>
 #include <cradle/inner/service/resources.h>
+#include <cradle/thinknode/disk_cached.h>
 
 namespace cradle {
+
+// This is legacy code, not using the requests architecture.
 
 // This function is a coroutine so takes `key` by value.
 // The `cache` lifetime is assumed to exceed the coroutine's one.
@@ -76,6 +78,64 @@ fully_cached(
             return disk_cached<Value>(
                 resources, key1, std::move(task_creator));
         });
+}
+
+namespace detail {
+
+/*
+ * Coroutine co_await'ing a shared task and tracking this
+ *
+ * cache_key must be available after the initial suspension point, so ownership
+ * must be inside this function.
+ */
+template<typename Value>
+cppcoro::shared_task<Value>
+shared_task_wrapper(
+    cppcoro::shared_task<Value> shared_task,
+    tasklet_tracker* client,
+    captured_id cache_key,
+    std::string summary)
+{
+    client->on_before_await(summary, *cache_key);
+    auto res = co_await shared_task;
+    client->on_after_await();
+    co_return res;
+}
+
+} // namespace detail
+
+/**
+ * Makes a shared task producing some cacheable object, on behalf of a tasklet
+ * client
+ *
+ * - Is or wraps a cppcoro::shared_task<Value> object.
+ * - The cacheable object is identified by a captured_id.
+ * - client will be nullptr while introspection is disabled.
+ *
+ * This construct has to be used when needing to co_await on a coroutine that
+ * calculates the cache key. If co_await and key calculation are co-located, a
+ * direct tasklet_await is also possible. (Both options are currently in use.)
+ */
+template<typename Value, typename TaskCreator>
+cppcoro::shared_task<Value>
+make_shared_task_for_cacheable(
+    inner_resources& resources,
+    captured_id const& cache_key,
+    TaskCreator task_creator,
+    tasklet_tracker* client,
+    std::string summary)
+{
+    auto shared_task
+        = fully_cached<Value>(resources, cache_key, std::move(task_creator));
+    if (client)
+    {
+        return detail::shared_task_wrapper<Value>(
+            std::move(shared_task), client, cache_key, std::move(summary));
+    }
+    else
+    {
+        return shared_task;
+    }
 }
 
 } // namespace cradle
