+++ src/cradle/thinknode/iam.cpp
@@ -1,10 +1,9 @@
 #include <cradle/thinknode/iam.h>
 
 #include <cradle/inner/core/sha256_hash_id.h>
-#include <cradle/inner/introspection/tasklet.h>
+#include <cradle/thinknode/caching.h>
 #include <cradle/thinknode/disk_cache_serialization.h>
 #include <cradle/thinknode/utilities.h>
-#include <cradle/typing/core/monitoring.h>
 #include <cradle/typing/io/http_requests.hpp>
 
 namespace cradle {
