+++ src/cradle/websocket/local_calcs.cpp
@@ -17,6 +17,7 @@
 #include <cradle/inner/fs/file_io.h>
 #include <cradle/inner/utilities/errors.h>
 #include <cradle/inner/utilities/functional.h>
+#include <cradle/thinknode/caching.h>
 #include <cradle/thinknode/supervisor.h>
 #include <cradle/thinknode/utilities.h>
 #include <cradle/typing/core/dynamic.h>
